# src/middleware/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: REST API for Users and Posts Database
  version: 1.0
paths:
...
  /api/subscriptions/public/user:
    get:
      summary: Get all user subscriptions by user id
      tags:
        - subscriptions
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of all user subscriptions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '400':
          description: Bad request
        '401':
          description: Unauthorized - you must sign in
        '404':
          description: Subscriptions are not found
        '422':
          description: Unprocessable entity
        '500':
          description: Internal server error
  /api/subscriptions/public/user/{subscriptionId}:
    get:
      summary: Get a user subscription by id
      tags:
        - subscriptions
      security:
        - BearerAuth: []
      parameters:
        - name: subscriptionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: |
            Subscription ID. Example: `6dcd8f9a-3b7c-4e5f-ba9e-2f8a7f8f8f8f`
      responses:
        '200':
          description: User subscription details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '400':
          description: Bad request
        '401':
          description: Unauthorized - you must sign in
        '404':
          description: Subscriptions are not found
        '422':
          description: Unprocessable entity
        '500':
          description: Internal server error
    delete:
      summary: Delete a user subscription by id
      tags:
        - subscriptions
      security:
        - BearerAuth: []
      parameters:
        - name: subscriptionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: |
            Subscription ID. Example: `6dcd8f9a-3b7c-4e5f-ba9e-2f8a7f8f8f8f`
      responses:
        '200':
          description: Subscription successfully deleted
        '400':
          description: Bad request
        '401':
          description: Unauthorized - you must sign in
        '404':
          description: Subscription is not found
        '422':
          description: Unprocessable entity
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Subscription:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        period:
          type: string
          format: date-time
        userName:
          type: string
        subscriberName:
          type: string
        userId:
          type: string
          format: uuid
        subscriberId:
          type: string
          format: uuid
...